+++ src/components/FormElements/FormError/styles.ts
@@ -1,11 +1,10 @@
 import { css, Theme, SerializedStyles } from "@emotion/react";
-import { hexToRGBA } from "@theme/default/colors";
 
 export const errorContainer = ({ typeScaleSizes, formElements }: Theme): SerializedStyles => css`
   font-size: ${typeScaleSizes.md};
   width: 100%;
   display: inline-block;
-  background-color: ${hexToRGBA(formElements.generalError.background, 0.25)};
+  background-color: ${formElements.generalError.background};
   border: 1px solid ${formElements.generalError.borderColor};
   border-radius: 2px;
   padding: 1.5rem 2rem;