+++ src/theme/default/config/badge.ts
@@ -3,4 +3,5 @@ import { colors } from "../colors";
 export default {
   background: colors.red.base,
   color: colors.white,
+  border: colors.white,
 };