+++ src/theme/default/config/drawer.ts
@@ -2,7 +2,8 @@ import { colors } from "../colors";
 
 export default {
   closeBtnColor: {
-    base: colors.grayBlue.base,
-    hover: colors.grayBlue.darkest,
+    base: colors.black,
+    hover: colors.secondary.base,
   },
+  dialog: colors.white,
 };