+++ src/theme/default/config/modal.ts
@@ -1,7 +1,8 @@
 import { colors } from "../colors";
 
 export default {
-  iconHover: colors.primary.base,
+  iconHover: colors.secondary.base,
   color: colors.black,
-  border: colors.grayBlue.light,
+  border: colors.secondary.lighter,
+  content: colors.white,
 };