+++ src/theme/default/config/progressBar.ts
@@ -1,7 +1,7 @@
 import { colors } from "../colors";
 
 export default {
-  background: "#E6E6E6",
+  background: colors.secondary.lighter,
   progressBackground: colors.green.base,
   percentColor: colors.white,
 };