+++ src/theme/default/config/result.ts
@@ -2,5 +2,5 @@ import { colors } from "../colors";
 
 export default {
   titleColor: colors.black,
-  infoColor: colors.gray.base,
+  infoColor: colors.secondary.base,
 };