+++ src/theme/default/config/tabs.ts
@@ -2,6 +2,6 @@ import { colors } from "../colors";
 
 export default {
   headerBackground: colors.white,
-  headerBorder: "#ccc",
+  headerBorder: colors.secondary.base,
   linkBorder: colors.primary.base,
 };