+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+## [3.5.5](https://github.com/epignosis/gnosis/compare/v3.5.4...v3.5.5) (2023-05-22)
+
+## [3.5.4](https://github.com/epignosis/gnosis/compare/v3.5.3...v3.5.4) (2023-05-22)
+
+## [3.5.3](https://github.com/epignosis/gnosis/compare/v3.5.2...v3.5.3) (2023-05-19)
+
+
+### Bug Fixes
+
+* updating color pallette ([#145](https://github.com/epignosis/gnosis/issues/145)) ([976e89d](https://github.com/epignosis/gnosis/commit/976e89d69e56b12fa1fd7e99ca61383e7040b04a))
+
+## [3.5.2](https://github.com/epignosis/gnosis/compare/v3.5.1...v3.5.2) (2023-05-18)
+
+
+### Bug Fixes
+
+* dummy comment to release changes ([#153](https://github.com/epignosis/gnosis/issues/153)) ([54d838e](https://github.com/epignosis/gnosis/commit/54d838e667173a2d0adcee609518ee890af0181a))
+
+## [3.5.1](https://github.com/epignosis/gnosis/compare/v3.5.0...v3.5.1) (2023-05-03)
+
+# [3.5.0](https://github.com/epignosis/gnosis/compare/v3.4.3...v3.5.0) (2023-05-02)
+
+
+### Features
+
+* **table.tsx:** Add a callback for hover actions ([#150](https://github.com/epignosis/gnosis/issues/150)) ([d55f0ee](https://github.com/epignosis/gnosis/commit/d55f0ee8046d4002e81ae84dbd4ec1f05d8a74a3))
+
+## [3.4.3](https://github.com/epignosis/gnosis/compare/v3.4.2...v3.4.3) (2023-04-26)
+
+## [3.4.2](https://github.com/epignosis/gnosis/compare/v3.4.1...v3.4.2) (2023-04-24)
+
+## [3.4.1](https://github.com/epignosis/gnosis/compare/v3.4.0...v3.4.1) (2023-04-19)
+
+# [3.4.0](https://github.com/epignosis/gnosis/compare/v3.3.3...v3.4.0) (2023-04-03)
+
+
+### Features
+
+* introducing table component in gnosis ([#143](https://github.com/epignosis/gnosis/issues/143)) ([d3625eb](https://github.com/epignosis/gnosis/commit/d3625eb4cfd246b2a239b77723ddae4c2c639081))
+
 ## [3.3.3](https://github.com/epignosis/gnosis/compare/v3.3.2...v3.3.3) (2023-03-16)
 
 ## [3.3.2](https://github.com/epignosis/gnosis/compare/v3.3.1...v3.3.2) (2023-03-14)